services/news-aggregator/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
apscheduler==3.10.4
feedparser==6.0.11
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
asyncpg==0.30.0
redis[asyncio]==5.2.1
scikit-learn==1.5.2
numpy==2.1.3
pydantic-settings==2.6.1
python-dateutil==2.9.0
